requirements.txt
absl-py==2.4.0
attrs==26.1.0
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.1
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.14
ml_dtypes==0.5.4
numpy==1.26.4
opencv-contrib-python==4.8.1.78
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pillow==12.1.1
protobuf==4.23.4
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.5
